requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
flask-cors==5.0.1
python-dotenv==1.1.0